Extracted public C and C++ interfaces implementation in persistence
