+++ modules/core/src/persistence_yml.cpp
@@ -1,3 +1,8 @@
+// This file is part of OpenCV project.
+// It is subject to the license terms in the LICENSE file found in the top-level directory
+// of this distribution and at http://opencv.org/license.html
+
+
 #include "precomp.hpp"
 #include "persistence.hpp"
